fix(ingress): move redirect logic to redirect.go
* sum all redirect logic
* refactor to use source instead of app in redirect code
