+++ pkg/generators/application.go
@@ -6,7 +6,6 @@ import (
 
 	nais_io_v1 "github.com/nais/liberator/pkg/apis/nais.io/v1"
 	nais_io_v1alpha1 "github.com/nais/liberator/pkg/apis/nais.io/v1alpha1"
-	"github.com/nais/naiserator/pkg/redirect"
 	"github.com/nais/naiserator/pkg/resourcecreator/frontend"
 	"github.com/nais/naiserator/pkg/resourcecreator/login"
 	"github.com/nais/naiserator/pkg/resourcecreator/observability"
@@ -90,7 +89,7 @@ func (g *Application) Prepare(ctx context.Context, source resource.Source, kube 
 	}
 
 	// Check if the application is allowed to redirect to another host
-	if err = redirect.Allowed(ctx, app, kube); err != nil {
+	if err = ingress.RedirectAllowed(ctx, app, kube); err != nil {
 		return nil, err
 	}
